% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoReport.R
\name{autoReport}
\alias{autoReport}
\alias{subscriptionLocalMonthlyReps}
\alias{dispatchMonthyKi}
\alias{dispatchMonthlyKi}
\title{Make sheduled automated reports}
\usage{
dispatchMonthlyKi(
  baseName,
  hospitalName,
  reshID,
  author,
  userRole,
  type,
  registryName
)

subscriptionLocalMonthlyReps(
  baseName,
  reshId,
  registryName,
  author,
  hospitalName,
  type
)
}
\arguments{
\item{baseName}{Single element list with value of Rmd template for the
report.
Value must be provided without the file extention (\emph{i.e.} ".Rmd")}

\item{hospitalName}{Single element list with a (human readable) org name as
its value}

\item{reshID}{Single element list with the organization id of the subscriber
as its value}

\item{author}{Single element list holding the name of the subscriber as its
value}

\item{userRole}{String defining the user role}

\item{type}{Single element list which value defining report file format,
currently one of "pdf" or "html"}

\item{registryName}{Single element list which value is the regisry name}

\item{reshId}{Single element list with the organization id of the subscriber
as its value}
}
\value{
Full path of file produced
}
\description{
When called, typically through a chain of events initiated by a chron
process, these are functions that actually produces the reports to be
shipped off to the recipients. The actual call to this function is made
through do.call and has the effect of providing the parameters as class
\emph{list}. Hence, values must be extracted by list operations
}
