% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{avdod_prosedyrerelatert_opphold}
\alias{avdod_prosedyrerelatert_opphold}
\title{Minst en gang prosedyrerelatert død under NORIC-oppholdet}
\usage{
avdod_prosedyrerelatert_opphold(df_ap)
}
\arguments{
\item{df_ap}{data.frame med AngioPCI-data tabellen. Må inneholde variablene 
\code{AvdRESH}, \code{OppholdsID} og \code{dod_prosedyrerelatert_noric}.}
}
\value{
Funksjonen returnerer \code{ap_df}, med en ny kolonne
ved navn \code{dod_prosedyrerelatert_opphold}.
}
\description{
Funksjonen grupperer primær- og sekundærforløp sammen, ved hjelp 
av \code{OppholdsID}. En ny variabel blir generert i AP-tabellen: 
\code{dod_prosedyrerelatert_opphold}. Denne variabelen får verdien \emph{Ja} dersom  
\code{dod_prosedyrerelatert_noric} har verdien \emph{Ja} for minst et av forløpene med felles
\code{OppholdsID} og får verdien \emph{Nei} dersom ingen av forløpene 
med felles \code{OppholdsID} har registrering av avdød. Funksjonen brukes
sammen med \code{noric::utlede_OppholdsID} og 
\code{noric::utlede_dod_prosedyrerelatert_noric}.
}
\examples{
x <- data.frame(AvdRESH = rep(123456, 8), 
               ForlopsID = 1:8, 
               Regtype = c("Primær", "Sekundær", rep("Primær", 4), 
                           "Sekundær", "Sekundær"), 
               PrimaerForlopsID = c(1, 1, 2, 2, 3, 4, 4, 4), 
               LabKompProsedyrerelatertDod = c("Nei", "Nei", "Ukjent", NA, 
                                               "Ja", "Nei", "Ukjent", "Ja"), 
               AvdKompProsedyrerelatertDod = c("Nei","Ja", 
                                               "Nei", "Nei", 
                                               "Ukjent", "Ukjent", 
                                                NA, NA))
x \%>\%  
  noric::utlede_OppholdsID(.) \%>\% 
  noric::utlede_dod_prosedyrerelatert_noric(.) \%>\% 
  noric::avdod_prosedyrerelatert_opphold(.)  \%>\% 
  dplyr::arrange(OppholdsID)              


}
