% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erstatt_koder_m_etiketter.R
\name{erstatt_koder_m_etiketter}
\alias{erstatt_koder_m_etiketter}
\title{Erstatt kode med etiketter i data.frame basert på kodebok}
\usage{
erstatt_koder_m_etiketter(df)
}
\arguments{
\item{df}{data.frame, er kun kvalitetssikret til å fungere med angiopcinum}
}
\value{
data.frame med etiketter i stedet for koder
}
\description{
Erstatt kode med etiketter i data.frame basert på kodebok
}
