% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fjerne_tulleregistreringer.R
\name{fjerne_tulleregistreringer}
\alias{fjerne_tulleregistreringer}
\title{Remove test-registrations}
\usage{
fjerne_tulleregistreringer(df, var = ProsedyreDato)
}
\arguments{
\item{df}{One of the NORIC-tables: AP, KA, FO etc (Must contain variable
`AvdRESH`)}

\item{var}{variable name for date of procedure, used for filtering old
procedures. AP, anP, anD og SS uses `ProsedyreDato`, CT uses `undersokDato`
FO og SO uses `HovedDato`}
}
\value{
data.frame where old procedures are removed
}
\description{
Remove registrations done before the hospitals officially joined NORIC
}
\examples{
x <- data.frame(AvdRESH = c(rep(102966,3), rep(109880,3)),
                ProsedyreDato = rep(c("2012-01-31",
                                      "2013-02-28",
                                      "2020-06-19"),2))
}
