% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAKData.R
\name{getAKData}
\alias{getAKData}
\title{getAKData provides local or national reg data from AortaklaffVar}
\usage{
getAKData(registryName, singleRow = FALSE, ...)
}
\arguments{
\item{registryName}{String providing the registry name}

\item{singleRow}{Logical defining if only one row is to be returned. A
relevant usecase will be when only description is needed. By default set to
FALSE}

\item{...}{Optional arguments to be passed to the function}
}
\value{
Data frame representing the table AortaklaffVar
}
\description{
getAKData provides local or national reg data from AortaklaffVar
}
