% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAKOppfData.R
\name{getAKOppfData}
\alias{getAKOppfData}
\title{getAKOppfData provides local reg data from AortaklaffOppfVar}
\usage{
getAKOppfData(registryName, singleRow = FALSE, ...)
}
\arguments{
\item{registryName}{String providing the registry name}

\item{singleRow}{Logical defining if only one row is to be returned. A
relevant usecase will be when only description is needed. By default set to
FALSE}

\item{...}{Optional arguments to be passed to the function}
}
\value{
Data frame representing the table AortaklaffOppfVar
}
\description{
getAKOppfData provides local reg data from AortaklaffOppfVar
}
