% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPDataLight.R
\name{getAPDataLight}
\alias{getAPDataLight}
\title{getAPDataLight provides local reg data from AngioPCIVar + useful generated
variables}
\usage{
getAPDataLight(registryName, singleRow = FALSE, ...)
}
\arguments{
\item{registryName}{String providing the registry name}

\item{singleRow}{Logical defining if only one row is to be returned. A
relevant usecase will be when only description is needed. By default set to
FALSE}

\item{...}{Optional arguments to be passed to the function}
}
\value{
Data frame representing a simplified version of angio PCI-data
}
\description{
getAPDataLight provides local reg data from AngioPCIVar + useful generated
variables
}
