% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataDump.R
\name{getDataDump}
\alias{getDataDump}
\title{Query data for dumps}
\usage{
getDataDump(registryName, tableName, fromDate, toDate, singleHospital, ...)
}
\arguments{
\item{registryName}{String registry name}

\item{tableName}{String with name of table to query from}

\item{fromDate}{Character string of format YYYY-MM-DD with start date. Value
NULL if no filter on date.}

\item{toDate}{Character string of format YYYY-MM-DD with end date. Value
NULL if no filter on date.}

\item{singleHospital}{Integer. Contains reshID from which data is loaded. 
Value 0 is national.}

\item{...}{Additional parameters to be passed to the function}
}
\value{
A data frame with registry data
}
\description{
Query data for dumps
}
