% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKodebokData.R
\name{getKodebokData}
\alias{getKodebokData}
\title{Kodebok}
\usage{
getKodebokData()
}
\value{
kodebok for taviprom (føsrte versjon)
}
\description{
Hent kodebok (metadata) for alle NORIC variabler. Merk: Kun tavi prom i 
første versjon.
}
\examples{
noric::getKodebokData() \%>\% 
  dplyr::filter(fysisk_feltnavn == "heart01")
}
