% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKodebokData.R
\name{getKodebokMedUtledeteVariabler}
\alias{getKodebokMedUtledeteVariabler}
\title{Kodebok med utledete variabler}
\usage{
getKodebokMedUtledeteVariabler()
}
\value{
data.frame med kodebok for NORIC med utledete variabler
}
\description{
Legg til utledete variabler i kodebok for NORIC
}
