% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLatestEntry.R
\name{getLatesEntry}
\alias{getLatesEntry}
\title{Get the latest registration date from the registry database}
\usage{
getLatesEntry(registryName)
}
\arguments{
\item{registryName}{}
}
\value{
A Date object
}
\description{
Provides the newest date based on HovedDato from SkjemaOversikt
}
