% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocalAnPData.R
\name{getLocalAnPData}
\alias{getLocalAnPData}
\title{getLocalAnPData provides local reg data from AndreProsedyrerVar}
\usage{
getLocalAnPData(registryName, singleRow = FALSE, ...)
}
\arguments{
\item{registryName}{String providing the registry name}

\item{singleRow}{Logical defining if only one row is to be returned. A
relevant usecase will be when only description is needed. By default set to
FALSE}

\item{...}{Optional arguments to be passed to the function}
}
\value{
Data frame representing the table AndreProsedyrerVar
}
\description{
getLocalAnPData provides local reg data from AndreProsedyrerVar
}
