% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocalAPData.R
\name{getLocalAPData}
\alias{getLocalAPData}
\title{getLocalAPData provides local reg data from AngioPCIVar}
\usage{
getLocalAPData(registryName, singleRow = FALSE, ...)
}
\arguments{
\item{registryName}{String providing the registry name}

\item{singleRow}{Logical defining if only one row is to be returned. A
relevant usecase will be when only description is needed. By default set to
FALSE}

\item{...}{Optional arguments to be passed to the function}
}
\value{
Data frame representing the table AngioPCIVar
}
\description{
getLocalAPData provides local reg data from AngioPCIVar
}
