% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMKData.R
\name{getMKData}
\alias{getMKData}
\title{getMKData provides local or national reg data from MitralklaffVar}
\usage{
getMKData(registryName, singleRow = FALSE, ...)
}
\arguments{
\item{registryName}{String providing the registry name}

\item{singleRow}{Logical defining if only one row is to be returned. A
relevant usecase will be when only description is needed. By default set to
FALSE}

\item{...}{Optional arguments to be passed to the function}
}
\value{
Data frame representing the table MitralklaffVar
}
\description{
getMKData provides local or national reg data from MitralklaffVar
}
