% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_segmentStent_variabler.R
\name{legg_til_antall_stent}
\alias{legg_til_antall_stent}
\title{Add number of stents per procedure in AP-dataset}
\usage{
legg_til_antall_stent(ap, ss)
}
\arguments{
\item{ap}{Angio PCI dataset where new variable `antall_stent` should be
added. Must contain the variables `AvdRESH` and `ForlopsID`,}

\item{ss}{segment stent dataset where `antall_stent` is calculated from.
Must contain the variables `AvdRESH`, `ForlopsID` and `StentType`}
}
\value{
data.frame with one new variable
}
\description{
Add number of stents per procedure in AP-dataset
}
\examples{
\dontrun{
AP \%>\% legg_til_antall_stent(., ss = SS)
}
}
