% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_ventetider.R
\name{legg_til_liggedogn}
\alias{legg_til_liggedogn}
\title{Liggedogn}
\usage{
legg_til_liggedogn(df_ap)
}
\arguments{
\item{df_ap}{data.frame med Angio PCI data fra noric. Maa inneholde
variablene \code{AnkomstPCIDato}, \code{UtskrivingsDato},
\code{OverflyttetFra} og \code{Regtype}.}
}
\value{
Returnerer \code{df_ap} med to nye variabler
 \code{liggedogn} og \code{liggedogn_dg}. Variabelen \code{liggedogn}
 inneholder antall dager mellom AnkomstPCI-sykehus og utskrivelse, denne
 kan ogsaa inneholde negative tider og tider over 1 aar.
 Variabelne \code{liggetid_dg} inneholder datagrunnlaget for
 \code{liggedogn} og har verdiene:
\itemize{
\item "nei" for pasienter overfoert fra annen avdeling paa sykehuset eller
der denne informasjonen er manglende.
\item "nei" for sekundaerforlop, da disse ikke har utskrivelsesskjema.
\item "ja" for primaerforlop hos direkte innlagte pasienter eller pasienter
overfoert fra andre sykehus og der liggedogn er i intervallet 0 dager
til 1 aar.
\item "ugyldig" for primaerforlop hos direkte innlagte pasienter
eller pasienter overfoert fra andre sykehusmed og der liggetid er enten et
negativt antall dager eller over 1 aar.
\item "manglende" for primaerforlop hos direkte innlagte pasienter eller
pasienter overfoert fra andre sykehus der utskrivingsdato og/eller
AnkomstPCIDato mangler.
}
}
\description{
Antall dager fra AnkomstPCIDato til Utskrivingsdato, kun for primaerforlop
hos direkte innlagte pasienter eller pasienter overfoert fra annet
sykehus.
}
