% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_ventetider.R
\name{legg_til_liggedogn_tavi}
\alias{legg_til_liggedogn_tavi}
\title{Liggedogn TAVI}
\usage{
legg_til_liggedogn_tavi(df_ak)
}
\arguments{
\item{af_ak}{data.frame med Aortaklaff data fra noric. Maa inneholde
variablene \code{ProsedyreDato}, \code{UtskrDato},
og \code{UtskrevetTil}.}
}
\value{
Returnerer \code{df_ak} med to nye variabler
 \code{liggedogn_tavi} og \code{liggedogn_tavi_data}. 
 Variabelen \code{liggedogn_tavi}
 inneholder antall dager mellom prosedyren og utskrivelsen, denne
 kan ogsaa inneholde negative tider og tider over 30 dager.
 Variabelen \code{liggetid_tavi_data} inneholder datagrunnlaget for
 \code{liggedogn_tavi} og har verdiene:
\itemize{
\item "ja" der liggedogn_tavi er i intervallet 0 dager til 30 dager.
\item "ugyldig" der liggetid_tavi er enten et
negativt antall dager eller over 30 dager.
\item "manglende" dersom dato mangler
}
}
\description{
Antall dager fra prosedyren til utskrivingsdato etter prosedyrer på 
aortaklaffen.
}
