% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legg_til_tidsvariabler.R
\name{legg_til_tidsvariabler}
\alias{legg_til_tidsvariabler}
\title{Add time variables in data frame
Year, month, quarter, week..}
\usage{
legg_til_tidsvariabler(df, var = ProsedyreDato)
}
\arguments{
\item{df}{data.frame where variables should be added}

\item{var}{variable in format "%Y-%m-%d".}
}
\value{
`df` data.frame with new variables added
}
\description{
Add time variables in data frame
Year, month, quarter, week..
}
\examples{
\dontrun{
# FOR AP, AK and CT-data:
AP \%<>\% legg_til_tidsvariabler(., var = ProsedyreDato)
AK \%<>\% legg_til_tidsvariabler(., var = ProsedyreDato)
CT \%<>\% legg_til_tidsvariabler(., var = HovedDatp)
}
}
