% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_ventetider.R
\name{legg_til_ventetid_nstemi_timer}
\alias{legg_til_ventetid_nstemi_timer}
\title{NSTEMI ventetid}
\usage{
legg_til_ventetid_nstemi_timer(df_ap)
}
\arguments{
\item{df_ap}{tabellen med AngioPCI data fra NORIC. Maa inneholde variablene
\code{OverflyttetFra, ProsedyreDato, ProsedyreTid,AnkomstPCIDato,
 AnkomstPCITid, InnleggelseHenvisendeSykehusDato,
 InnleggelseHenvisendeSykehusTid}}
}
\value{
returnerer \code{df_ap} med en ny kolonne:
 \code{ventetid_nstemi_timer}
}
\description{
Antall timer mellom Ankomst PCI og Prosedyre for direkte innlagte pasienter.
Antall timer mellom Innleggelse ved henvisende sykeshus og Prosedyre for
overfoerte pasienter. Dersom en av datoene/tidspunktene mangler kan ikke
antall timer regnes ut. Dersom  \code{OverflyttetFra} mangler eller er
lik "Annen avd paa sykehuset" regnes ikke ventetid ut. Ingen filter paa
unormalt lange eller negative, ventetider.
}
\examples{
x <- data.frame(
 OverflyttetFra = c("Omdirigert ambulanse",
                    "Nei, direkte inn til dette sykehuset",
                    "Annet sykehus", NA),
 ProsedyreDato = as.Date(c("2020-01-30", "2021-11-15",
                           "2020-11-11", "2021-12-24"),
                         format = "\%Y-\%m-\%d"),

 ProsedyreTid = c("22:30:00", "01:10:00", "13:45:00", "12:20:00"),
   AnkomstPCIDato = as.Date(
   c("2020-01-28", "2021-11-14", "2018-04-24", "2020-01-01"),
   format = "\%Y-\%m-\%d"),
 AnkomstPCITid = c(
   "21:10:00", "23:10:00", "05:20:00", NA_character_),
 InnleggelseHenvisendeSykehusDato = as.Date(
   c(NA, NA,  "2020-11-10", "2021-12-21"),
   format = "\%Y-\%m-\%d"),
 InnleggelseHenvisendeSykehusTid = c(NA_character_, NA_character_,
                                     "11:25:00", "23:20:00"))
noric::legg_til_ventetid_nstemi_timer(x)





}
