% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_ventetider.R
\name{legg_til_ventetid_stemi_min}
\alias{legg_til_ventetid_stemi_min}
\title{STEMI ventetid}
\usage{
legg_til_ventetid_stemi_min(df_ap)
}
\arguments{
\item{df_ap}{tabellen med AngioPCI data fra NORIC. Maa inneholde variablene
\code{ProsedyreDato, ProsedyreTid,BesUtlEKGDato, BesUtlEKGTid}}
}
\value{
returnerer \code{df_ap} med en ny kolonne:
 \code{ventetid_stemi_minutter}
}
\description{
Antall minutter fra beslutningsutløsende EKG til arteriepunksjon.
Dersom en av datoene/tidspunktene mangler kan ikke ventetiden regnes ut.
Ingen filter paa unormalt lange eller negative, ventetider. Dette er en
hjelpefunksjon til \link[noric]{ki_stemi_pci_innen120min}, ventetidene
må ikke analyseres utenfor denne funksjonen da datagrunnlaget defineres der.
}
\examples{
x <- data.frame(
 ProsedyreDato = as.Date(c("2020-01-30", "2021-11-15",
                           "2020-11-11", "2021-12-24"),
                         format = "\%Y-\%m-\%d"),
 ProsedyreTid = c("22:30:00", "01:10:00", "13:45:00", "12:20:00"),
 BesUtlEKGDato = as.Date(c("2020-01-30", "2021-11-14",
                           "2018-04-24", "2020-01-01"),
                         format = "\%Y-\%m-\%d"),
 BesUtlEKGTid = c( "21:10:00", "23:10:00", "05:20:00", NA_character_),
 BeslutningsutlosendeEKG = rep("Ved annet sykehus", 4))
noric::legg_til_ventetid_stemi_min(x)
}
