% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_ventetider.R
\name{legg_til_ventetid_tavi}
\alias{legg_til_ventetid_tavi}
\title{Ventetid TAVI}
\usage{
legg_til_ventetid_tavi(df_ak)
}
\arguments{
\item{df_ak}{data.frame med AortaklaffVar data fra noric. Maa inneholde
variablene \code{BeslutningsDato} og \code{ProsedyreDato}}
}
\value{
Returnerer \code{df_ak} med en ny variabel
 \code{ventetid_tavi}. Variabelen \code{ventetid_tavi}
 inneholder antall dager mellom dato for beslutning og dato for
 prosedyre. Denne kan ogsaa inneholde negative tider og har ingen begrensing 
 for øvre ventetid.
}
\description{
Antall dager fra BeslutningsDato til ProsedyreDato for alle forløp på
Aortaklaffen.
}
