% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRegData.R
\name{loadRegDataAngioPCI}
\alias{loadRegDataAngioPCI}
\title{Provide dataframe for AngioPCI}
\usage{
loadRegDataAngioPCI(registryName)
}
\arguments{
\item{registryName}{String providing the current registryName}
}
\value{
RegData data frame
}
\description{
Provides a dataframe for AngioPCI from NORIC staging db. To be used in
NORIC_local_monthly_stent.Rnw and maybe others
}
