% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRegData.R
\name{loadRegDataStent}
\alias{loadRegDataStent}
\title{Provide dataframe for stents}
\usage{
loadRegDataStent(registryName)
}
\arguments{
\item{registryName}{String providing the current registryName}
}
\value{
RegData data frame
}
\description{
Provides a dataframe for stents from NORIC staging db. To be used in
NORIC_local_monthly_stent.Rnw and maybe others
}
