% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sampleDataGenerator.R
\name{makeAngioPCISampleData}
\alias{makeAngioPCISampleData}
\title{Sample data generator for AngioPCI}
\usage{
makeAngioPCISampleData(baseName, reshID)
}
\arguments{
\item{baseName}{String giving the prefix base of the name}

\item{reshID}{String providing the current reshID. At Rapporteket, reshID
should already be present in the current R session}
}
\value{
AngioPCI a data frame holding the data set
}
\description{
Functions to generate sample data for development purposes. These data
should be distributed openly, hence it must not hold any restricted
or sensitive information.
}
\details{
Resulting data can be saved by:
save([data_frame_name], file="[filename].rda")
}

