% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NORICloadRegData.R
\name{NORICloadRegDataAngioPCI}
\alias{NORICloadRegDataAngioPCI}
\title{Provide dataframe for AngioPCI}
\usage{
NORICloadRegDataAngioPCI(registryName)
}
\arguments{
\item{registryName}{String providing the current registryName}
}
\value{
RegData data frame
}
\description{
Provides a dataframe for AngioPCI from NORIC staging db. To be used in
NORIC_local_monthly_stent.Rnw and maybe others
}

