% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NORICloadRegData.R
\name{NORICloadRegDataStent}
\alias{NORICloadRegDataStent}
\title{Provide dataframe for stents}
\usage{
NORICloadRegDataStent(registryName)
}
\arguments{
\item{registryName}{String providing the current registryName}
}
\value{
RegData data frame
}
\description{
Provides a dataframe for stents from NORIC staging db. To be used in
NORIC_local_monthly_stent.Rnw and maybe others
}
