% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prom_funksjoner.R
\name{prom_funksjoner}
\alias{prom_funksjoner}
\alias{kriterie_taviprom_siste_aar}
\alias{kriterie_pciprom_siste_aar}
\alias{kriterie_ingen_ny_tavi}
\title{NORIC PCI prom}
\usage{
kriterie_taviprom_siste_aar(df, registryName = NULL)

kriterie_pciprom_siste_aar(df, registryName = NULL)

kriterie_ingen_ny_tavi(df, registryName = NULL)
}
\arguments{
\item{df}{pci + pciprom}

\item{registryName}{for SQL queries}
}
\description{
Hjelpefunksjoner for PCI prom
}
