% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{slaa_sammen_variabler}
\alias{slaa_sammen_variabler}
\title{Use values from two variables}
\usage{
slaa_sammen_variabler(df, var1, var2, var_name, slette_gamle = FALSE)
}
\arguments{
\item{df}{data.frame, must have variables \code{var1} and \code{var2}.}

\item{var1}{variable to be used as main-source for \code{var_name}}

\item{var2}{variable to use when rows of \code{var1} are missing}

\item{var_name}{string with name of new variable}

\item{slette_gamle}{boolean. Should \code{var1} and \code{var2} be
deleted from \code{df} before \code{df} is returned? Default value is FALSE.}
}
\description{
The function \code{slaa_sammen_variabler()} uses the values of two variables,
\code{var1} and \code{var2}, in dataset \code{df} to create a new variable
named \code{var_name} in the same dataset. \code{var_name} is first given the
value of \code{var1}, in rows where \code{var1} is missing the values of
\code{var2} are used. In rows where both \code{var1} and \code{var2} are
missing, \code{var_name} is also missing. If \code{slette_var1_var2} is TRUE,
\code{var1} and \code{var2} are deleted before \code{df} is returned.
}
\examples{
x <- data.frame(var1 = c(rep("A", 5), NA, NA),
                var2 = c(NA, rep("B", 5), NA))
slaa_sammen_variabler(df = x, var1 = var1, var2 = var2,
                      var_name = "var_1_2", slette_gamle = FALSE)
}
