% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stagingDataFunctions.R
\name{stagingData}
\alias{makeStagingDataKi}
\alias{makeStagingDataFrame}
\alias{checkValidStagingData}
\alias{deleteOldStagingData}
\alias{bulletinProcessorStaging}
\title{NORIC staging data functions}
\usage{
makeStagingDataKi(registryName, rendered_by_shiny = FALSE)

makeStagingDataFrame(registryName)

checkValidStagingData(registryName, diffDaysCheck = 0)

deleteOldStagingData(registryName, diffDaysDelete)

bulletinProcessorStaging(
  dataset = "ki",
  orgName = "unknown organization",
  orgId = 999999,
  registryName = "noric",
  userFullName = "unknown person name",
  userRole = "unknown role",
  userOperator = "unknown operator",
  rendered_by_shiny = FALSE,
  author = "ingen"
)
}
\arguments{
\item{registryName}{Character string defining the registry name.}

\item{rendered_by_shiny}{boolean. if TRUE progression of pdf-generation is
returned.}

\item{diffDaysCheck}{numerical. Default values is 0 (today).
\code{checkValidStagingData()} checks if any staging data has been created
within the \code{diffDaysCheck} days.}

\item{diffDaysDelete}{numerical. No default value.
\code{deleteOldStagingData()} deletes staging data older than this.}

\item{dataset}{Which kind of staging data to create}

\item{orgName}{Character string with the name of the organization/hospital.
Default is "unknown organization".}

\item{orgId}{Integer (?) with the id of the organization/hospital. Default
is 999999.}

\item{userFullName}{Character string giving the person name, normally the
user requesting the report. Default is "unknown person name".}

\item{userRole}{Character string giving a user role, normally the one of the
user requesting the report. Default is "unknown role".}

\item{userOperator}{Character string with some name of an operator, whatever
that is... Default is "unknown operator".}

\item{author}{"ukjent"}
}
\description{
NORIC's handling of staging data at Rapporteket.
}
\details{
\itemize{
\item \code{makeStagingDataKi()} creates staging data for the KI-report.
\item \code{makeStagingDataFrame()} lists all of the existing staging data.
\item \code{checkValidStagingData()} checks whether any staging data set has
 been created within the last \code{diffDaysCheck} days. Returns
\code{valid_staging_data} is TRUE and the name of the most recently created
data set \code{nyeste_staging_data} if this is the case, and FALSE elsewhere.
\item \code{deleteOldStagingData()} deletes staging data older than
\code{diffDaysDelete}
\item \code{bulletinProcessorStaging()} is a modified version of
\code{reportProcessor()}. The function is applied to create bulletins for
staged data.
}
}
