% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tavi_funksjoner.R
\name{tavi_funksjoner}
\alias{tavi_funksjoner}
\alias{legg_til_taviStatus}
\alias{legg_til_taviErrorCode}
\title{NORIC tavi prom}
\usage{
legg_til_taviStatus(df)

legg_til_taviErrorCode(df)
}
\arguments{
\item{df}{AortaklaffVar tabell}
}
\value{
df, with new variables suffixed \emph{_tekst}.
}
\description{
Functions for adding "listetekst" for variables i \emph{TaviProm}-table.
}
\examples{
data.frame(ePromStatus  = 0:4) \%>\% noric::legg_til_taviStatus()

}
