% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_alder}
\alias{utlede_alder}
\title{Add variable PasientAlder}
\usage{
utlede_alder(df, var = ProsedyreDato)
}
\arguments{
\item{df}{data.frame, must contain a column with FodselsDato}

\item{var}{is the variable name of procedure date in \code{df}. Default
value is \code{ProcedureDate}.}
}
\value{
The input data.frame \code{df} is returned as it is,  with one new
 column named \code{PasientAlder} added.
}
\description{
Years between birth date and proceduredate (depends on df)
}
\examples{
df <-data.frame(FodselsDato = as.Date(c("1950-01-01", "1999-01-01")), 
                ProsedyreDato = as.Date(c("2020-01-10", "2024-01-10")))
utlede_alder(df) 
}
