% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_dod_noric}
\alias{utlede_dod_noric}
\title{Minst en registrering av død under forløpet}
\usage{
utlede_dod_noric(df_ap)
}
\arguments{
\item{df_ap}{data.frame med AngioPCI-data tabellen. Må inneholde variablene 
\code{LabKompDod}, \code{AvdKompDod}, \code{UtskrevetDod} og
\code{UtskrevetDodsdato}.}
}
\value{
Funksjonen returnerer \code{ap_df}, med en ny kolonne
ved navn \code{dod_noric}.
}
\description{
Funksjonen \code{utlede_dod_noric} genererer en ny variabel i AP-datasettet:
\code{dod_noric}. Dersom minst en av variablene 
\code{LabKompDod}, \code{AvdKompDod} eller \code{UtskrevetDod} har verdien 
\emph{Ja}, eller \code{UtskrevetDodsdato} finnes, så regnes pasienten som 
registrert død under forløpet. Gjelder alle typer forløp, primær og sekundær. 
Brukes sammen med funksjonen \code{avdod_opphold} for å gruppere 
NORIC-opphold sammen.
}
\examples{
x <- data.frame(LabKompDod = c("Ja", "Nei", "Ukjent"), 
                AvdKompDod = rep(NA_character_, 3), 
                UtskrevetDod = c("Nei", "Ja", "Nei"), 
                UtskrevetDodsdato = as.Date(c(NA_character_, 
                                              "2021-01-01", 
                                              NA_character_), 
                                              format = "\%Y-\%m-\%d"))
noric::utlede_dod_noric(x)                                               
}
