% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_dod_prosedyrerelatert_noric}
\alias{utlede_dod_prosedyrerelatert_noric}
\title{Minst en registrering av prosedyrerelatert død under forløpet}
\usage{
utlede_dod_prosedyrerelatert_noric(df_ap)
}
\arguments{
\item{df_ap}{data.frame med AngioPCI-data tabellen. Må inneholde variablene
\code{LabKompProsedyrerelatertDod} og \code{AvdKompProsedyrerelatertDod}.}
}
\value{
Funksjonen returnerer \code{ap_df}, med en ny kolonne
ved navn \code{dod_prosedyrerelatert_noric}.
}
\description{
Funksjonen \code{utlede_prosedyrerelatertdod_noric} genererer en 
ny variabel i AP-datasettet:
\code{dod_prosedyrerelatert_noric}. Dersom minst en av variablene 
\code{LabKompProsedyrerelatertDod} eller \code{AvdKompProsedyrerelatertDod} 
har verdien \emph{Ja},  så regnes pasienten som prosedyrerelatert død under 
forløpet. Gjelder alle typer forløp, primær og sekundær. 
Brukes sammen med funksjonen \code{avdod_prosedyrerelatert_opphold} for å gruppere 
NORIC-opphold sammen.
}
\examples{
x <- data.frame(LabKompProsedyrerelatertDod = c("Ja", "Nei", "Ukjent", NA, 
                                                "Ja", "Nei", "Ukjent", NA), 
                AvdKompProsedyrerelatertDod = c("Ja","Ja", 
                                                "Nei", "Nei", 
                                                "Ukjent", "Ukjent", 
                                                NA, NA))
noric::utlede_dod_prosedyrerelatert_noric(x)                                               
}
