% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_ferdigstilt}
\alias{utlede_ferdigstilt}
\title{Add binary variable for table-status}
\usage{
utlede_ferdigstilt(df, var = SkjemaStatusStart, suffix = "startSkjema")
}
\arguments{
\item{df}{data.frame, must contain a numerical status-variable with values
-1, 0 or 1.}

\item{var}{variable name of the numerical status-variable. Default value
is \code{SkjemaStatusStart}.}

\item{suffix}{wanted suffix in the new variable's name:
\code{ferdigstilt_suffix}. Default values is "startSkjema".}
}
\value{
The data.frame given in input \code{df} is returned as it is,
with one new column added.
}
\description{
The function \code{utlede_ferdigstilt()} uses the values of variable
\code{var} to create a new column named \code{ferdigstilt_suffix} in
\code{df} before \code{df} is returned from the function. The new variable is
given the value \emph{ja} if \code{var} is 1, and value \emph{nei} if
\code{var} is -1 or 0.
}
\examples{
x <- data.frame(SkjemaStatusStart = c(-1,-1,0,0,1,-1,0,1,NA),
                SkjemastatusHovedskjema = rep(0,9),
                SkjemaStatusUtskrivelse = rep(-1,9),
                SkjemaStatusKomplikasjoner = c(rep(-1,4), rep(1,3), NA, NA))

utlede_ferdigstilt(x, var = SkjemaStatusStart, suffix = "startSkjema")

utlede_ferdigstilt(x,
                   var = SkjemaStatusStart,
                   suffix = "startSkjema") \%>\%
   utlede_ferdigstilt(.,
                      var = SkjemastatusHovedskjema,
                      suffix = "hovedSkjema") \%>\%
   utlede_ferdigstilt(.,
                      var = SkjemaStatusKomplikasjoner,
                      suffix = "komplikSkjema")
}
