% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_segmentStent_variabler.R
\name{utlede_kar_graft_segment_stent}
\alias{utlede_kar_graft_segment_stent}
\title{Add variable `kar_graft` to NORIC segment-stent-table
Based on variables `Segment` and `Graft`. More detailed than
utlede_kar_segment_stent(), here also `Graft` is detailed for each level
of `kar`.}
\usage{
utlede_kar_graft_segment_stent(df = ss)
}
\arguments{
\item{df_ss}{segment-stent table, must contain variables `ForlopsID`,
`AVdRESH`, `Segment` and `Graft`}
}
\value{
segment-stent-table with one new variable
}
\description{
Add variable `kar_graft` to NORIC segment-stent-table
Based on variables `Segment` and `Graft`. More detailed than
utlede_kar_segment_stent(), here also `Graft` is detailed for each level
of `kar`.
}
\examples{
x <- data.frame(ForlopsID = 1:23,
                AvdRESH = rep(1,23),
                Segment = c(1:20, 1:3),
                Graft=c(rep("Nei", 20), "Arteriell", "Vene", NA))
x \%>\% utlede_kar_graft_segment_stent(.)
}
