% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_lab_avd_dod_noric}
\alias{utlede_lab_avd_dod_noric}
\title{Minst en registrering av død på lab/avdeling under forløpet}
\usage{
utlede_lab_avd_dod_noric(df_ap)
}
\arguments{
\item{df_ap}{data.frame med AngioPCI-data tabellen. Må inneholde variablene
\code{LabKompDod} og \code{AvdKompDod}.}
}
\value{
Funksjonen returnerer \code{ap_df}, med en ny kolonne
ved navn \code{dod_lab_avd_dod_noric}.
}
\description{
Funksjonen \code{utlede_lab_avd_dod_noric} genererer en 
ny variabel i AP-datasettet:
\code{dod_lab_avd_dod_noric}. Dersom minst en av variablene 
\code{LabKompDod} eller \code{AvdKompDod} 
har verdien \emph{Ja},  så regnes pasienten som  død under 
forløpet. Gjelder alle typer forløp, primær og sekundær. 
Brukes sammen med funksjonen \code{avdod_lab_avd_dod_noric_opphold} for å gruppere 
NORIC-opphold sammen.
}
\examples{
x <- data.frame(LabKompDod = c("Ja", "Nei", "Ukjent", NA, 
                               "Ja", "Nei", "Ukjent", NA), 
                AvdKompDod = c("Ja","Ja", 
                               "Nei", "Nei", 
                               "Ukjent", "Ukjent", 
                               NA, NA))
noric::utlede_lab_avd_dod_noric(x)                                               
}
