% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\alias{getAp}
\alias{getCt}
\alias{getAk}
\alias{getAkOppf}
\alias{getAnP}
\alias{getAnD}
\alias{getSs}
\alias{getMk}
\alias{getTaviProm}
\alias{getSh}
\alias{getFo}
\alias{getSo}
\alias{getPs}
\alias{getApLight}
\title{Load data for NORIC}
\usage{
getAp(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getCt(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getAk(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getAkOppf(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = NULL,
  ...
)

getAnP(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getAnD(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getSs(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getMk(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getTaviProm(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = NULL,
  ...
)

getFo(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getSo(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getPs(registryName, fromDate, toDate, singleRow, singleHospital = NULL, ...)

getApLight(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = NULL,
  ...
)
}
\arguments{
\item{registryName}{Character string defining the registry name.}

\item{fromDate}{Character string of format YYYY-MM-DD with start date. Value
NULL if no filter on date.}

\item{toDate}{Character string of format YYYY-MM-DD with end date. Value
NULL if no filter on date.}

\item{singleRow}{Logical if only one row from the table is to be provided.
Default value is FALSE.}

\item{singleHospital}{if only data from one hospital, when national database. 
Default value is NULL, contains reshID of selected hospital else.}

\item{...}{Optional arguments to be passed to the function.}
}
\value{
Data frame or (when multiple data sets are returned) a list of data
frames containing registry data. In case of \code{getNameReshId()} data may
also be returned as a named list of values (see Details).
}
\description{
Functions for loading data needed by NORIC from a database
}
\details{
The functions \code{getAp}, \code{getSs}, \code{getAk} etc. load the
AP-, SS- and AK- tables etc. respectively. For most tables, the query adds
selected variables from  \emph{forlopsoversikt}.
}
