% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataDump.R
\name{getDataDump}
\alias{getDataDump}
\title{Query data for dumps}
\usage{
getDataDump(registryName, tableName, fromDate, toDate, ...)
}
\arguments{
\item{registryName}{String registry name}

\item{tableName}{String with name of table to query from}

\item{fromDate}{String with start period, endpoint included}

\item{toDate}{String with end period, endpont included}

\item{...}{Additional parmeters to be passed to the function}
}
\value{
A data frame with registry data
}
\description{
Query data for dumps
}
