% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLatestEntry.R
\name{getLatestEntry}
\alias{getLatestEntry}
\title{Get the latest registration date from the registry database}
\usage{
getLatestEntry(registryName)
}
\arguments{
\item{registryName}{Character string providing registry name key}
}
\value{
A Date object
}
\description{
Provides the newest date based on HovedDato from SkjemaOversikt
}
