% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPivotDataSet.R
\name{getPivotDataSet}
\alias{getPivotDataSet}
\title{Get data set for Pivot-table}
\usage{
getPivotDataSet(setId = "", registryName, singleRow = FALSE, session, userRole)
}
\arguments{
\item{setId}{String defining name of data set to be returned}

\item{registryName}{String naming the registry data}

\item{singleRow}{Logical defining if only one row is to be returned. A
relevant usecase will be when only description/variable-list is needed.
By default set to FALSE}

\item{session}{List shiny session object}

\item{userRole}{String naming the user role}
}
\value{
data frame
}
\description{
Load pre-processed dataset into pivot table.
}
