% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndPrepareData.R
\name{getPrepData}
\alias{getPrepData}
\alias{getPrepApData}
\alias{getPrepCtData}
\alias{getPrepAkData}
\alias{getPrepAkOppfData}
\alias{getPrepAnPData}
\alias{getPrepAnDData}
\alias{getPrepSsData}
\alias{getPrepMkData}
\alias{getPrepTaviPromData}
\alias{getPrepFoData}
\alias{getPrepSoData}
\alias{getPrepPsData}
\alias{getPrepApLightData}
\title{Data managment on tables}
\usage{
getPrepApData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)

getPrepSoData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)

getPrepAkData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)

getPrepFoData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)

getPrepAnPData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)

getPrepCtData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)

getPrepAkOppfData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)

getPrepAnDData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)

getPrepSsData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)

getPrepMkData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)

getPrepPsData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)

getPrepApLightData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)

getPrepTaviPromData(
  registryName,
  fromDate,
  toDate,
  singleRow,
  singleHospital = 0,
  ...
)
}
\arguments{
\item{registryName}{String providing the registry name}

\item{fromDate}{A Date or character class object  of format "YYYY-MM-DD"
defining start of the period from which data will be collected. Default is
\code{NULL} in which case til will be set to the assumed pre-historic date of
"1900-01-01".}

\item{toDate}{A Date or character class object  of format "YYYY-MM-DD"
defining end of the period from which data will be collected. Default is
\code{NULL} in which case til will be set to the last known registration date
for NORIC.}

\item{singleRow}{Logical defining if only one row is to be returned. A
relevant usecase will be when only description is needed. By default set to
FALSE}

\item{singleHospital}{if only data from one hospital, when national database. 
Default value is 0, contains reshID of selected hospital else.}

\item{...}{Optional arguments to be passed to the function}
}
\value{
Data frame representing the chosen table. Basic data managment is
done (e.g. added variables from FO, added time-variables, rename hospitals)
}
\description{
Load data and apply data-managment operations. Tables can be used
in \emph{pivot-table} or in monthly reports.
}
