% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPresentHospitals.R
\name{getPresentHospitals}
\alias{getPresentHospitals}
\alias{getPresentHospitalsAp}
\alias{getPresentHospitalsAk}
\title{Get all hospitals present in the registry databases}
\usage{
getPresentHospitalsAp(registryName, fromDate = NULL, toDate = NULL)

getPresentHospitalsAk(registryName, fromDate = NULL, toDate = NULL)
}
\arguments{
\item{registryName}{Character string providing registry name key}

\item{fromDate}{Character string of format YYYY-MM-DD with start date. Value
NULL if no filter on date.}

\item{toDate}{Character string of format YYYY-MM-DD with end date. Value
NULL if no filter on date.#' @return A Date object}
}
\value{
character string with AvdRESH
}
\description{
Provides all AvdRESH present in AP and AK databases. Used to test
completness of data-transfert before creation of montly reports.
}
