% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuery.R
\name{getQuery}
\alias{getQuery}
\alias{queryAngiopcinum}
\alias{queryCtangiovarnum}
\alias{queryAortaklaffvarnum}
\alias{queryAortaklaffoppfvarnum}
\alias{queryAndreprosedyrervarnum}
\alias{queryAnnendiagnostikkvarnum}
\alias{querySegmentstentnum}
\alias{queryMitralklaffvarnum}
\alias{queryMitralklaffoppfvarnum}
\alias{queryTaviprom}
\alias{queryForlopsoversikt}
\alias{querySkjemaoversikt}
\alias{queryPasienterstudier}
\alias{queryApLight}
\alias{queryDiagnose}
\title{Fetch queries to NORIC raw data tables}
\usage{
queryAngiopcinum()

queryCtangiovarnum()

queryAortaklaffvarnum()

queryAortaklaffoppfvarnum()

queryAndreprosedyrervarnum()

queryAnnendiagnostikkvarnum()

querySegmentstentnum()

queryMitralklaffvarnum()

queryMitralklaffoppfvarnum()

queryTaviprom()

queryForlopsoversikt()

querySkjemaoversikt(fromDate, toDate, singleHospital)

queryPasienterstudier()

queryApLight()

queryDiagnose()
}
\arguments{
\item{fromDate}{Character string of format YYYY-MM-DD with start date. Value
NULL if no filter on date.}

\item{toDate}{Character string of format YYYY-MM-DD with end date. Value
NULL if no filter on date.}

\item{singleHospital}{NULL if national, reshid if query for one hospital}
}
\value{
query as string
}
\description{
Replace old views
}
