% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuery.R
\name{getQuery}
\alias{getQuery}
\alias{queryTaviprom}
\alias{queryAortaklaffvarnum}
\alias{queryForlopsoversikt}
\alias{queryAndreprosedyrervarnum}
\alias{queryAnnendiagnostikkvarnum}
\alias{querySegmentstentnum}
\alias{queryAngiopcinum}
\alias{queryMitralklaffvarnum}
\alias{queryMitralklaffoppfvarnum}
\title{Fetch queries to NORIC raw data tables}
\usage{
queryTaviprom()

queryAortaklaffvarnum()

queryForlopsoversikt()

queryAndreprosedyrervarnum()

queryAnnendiagnostikkvarnum()

querySegmentstentnum()

queryAngiopcinum()

queryMitralklaffvarnum()

queryMitralklaffoppfvarnum()
}
\value{
query as string
}
\description{
Replace old views
}
