% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legg_til_tidsvariabler.R
\name{legg_til_tidsvariabler}
\alias{legg_til_tidsvariabler}
\title{Add time-variables}
\usage{
legg_til_tidsvariabler(df, var = ProsedyreDato)
}
\arguments{
\item{df}{data.frame from any NORIC-table. Must contain a variable in
date-format.}

\item{var}{name of variable in date-format that should be used to calculate
the new variables.}
}
\value{
The input data.frame \code{df} is returned as it is, with six new
columns added,one for each time-variable.
}
\description{
The function \code{legg_til_tidsvariabler()} creates six time-variables,
\code{aar}, \code{maaned_nr}, \code{maaned}, \code{kvartal}, \code{uke},
\code{aar_uke}, based on a given date-column (\code{var}) of input
data.frame \code{df}. Then these new variables are added to \code{df} before
the entire data.frame is returned.
}
\examples{
df <- data.frame(
   datoVariabel = c("2021-02-28",
                    "2013-09-15",
                    "2020-05-31",
                    "2015-12-12",
                    "2016-12-31",
                    "2016-01-01" ))
legg_til_tidsvariabler(df = df,
                       var = datoVariabel)

\dontrun{
# FOR NORIC's ap, ak and ct-tables:
legg_til_tidsvariabler(ap, var = ProsedyreDato)
legg_til_tidsvariabler(ak, var = ProsedyreDato)
legg_til_tidsvariabler(ct, var = UndersokDato)
}
}
