% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapOrgReshId.R
\name{mapOrgReshId}
\alias{mapOrgReshId}
\title{Make mapping between unit name and id}
\usage{
mapOrgReshId(registryName, asNamedList = FALSE)
}
\arguments{
\item{registryName}{Character string providing registry name key}

\item{asNamedList}{Logical wether to return a list of named values or not.
Default is FALSE in which case a data frame containing name and id is
returned.}
}
\value{
A data frame with columns name and id or a list of named ids
}
\description{
Make mapping between unit name and id
}
