% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prom_funksjoner.R
\name{prom_funksjoner}
\alias{prom_funksjoner}
\alias{kriterie_taviprom_siste_aar}
\alias{kriterie_pciprom_siste_aar}
\alias{kriterie_ingen_tavi_neste3mnd}
\alias{legg_til_promStatus}
\alias{legg_til_promErrorCode}
\title{Hjelpefunksjoner for databehandling av PROM}
\usage{
kriterie_taviprom_siste_aar(df, registryName = NULL)

kriterie_pciprom_siste_aar(df, registryName = NULL)

kriterie_ingen_tavi_neste3mnd(df, registryName = NULL)

legg_til_promStatus(df)

legg_til_promErrorCode(df)
}
\arguments{
\item{df}{Inneholder data fra pci/tavi + respetiktive prom.

Noen df må inneholde variabelen \code{kriterie_alle_midlertidig}.}

\item{registryName}{for SQL queries}
}
\description{
Gjeler PCI-prom og TAVI-prom
}
\details{
For forløp der alle andre kriterier (alder, prosedyretype, indikasjon,
 utskrivelse, etc.) er oppfylt, det vil si
  \code{kriterie_alle_midlertidig} = \emph{ja}. 
 Sjekker om pasientene hadde eksisterende TAVI-PROM eller PCI-PROM det
 siste året. Sjekker om pasienten har fått TAVI i de 3 påfølgende månedene. 
 
 Legger til ny variabel: 
 
\code{kriterie_taviprom_siste_aar()}
 \itemize{
 \item \code{kriterie_ingen_taviprom har verdien} har verdien \emph{nei}
 dersom pasienten hadde TAVI-PROM det siste året.  
 }
\code{kriterie_pciprom_siste_aar()}
 \itemize{
 \item \code{kriterie_ingen_pciprom har verdien} har verdien \emph{nei}
 dersom pasienten hadde PCI-PROM det siste året.  
 }
 #' \code{kriterie_ingen_tavi_neste3mnd()}
 \itemize{
 \item \code{kriterie_ingen_ny_tavi har verdien} har verdien \emph{nei}
 dersom pasienten har vært til en TAVI-prosedyre i etterkant av PCI
  prosedyren (0-3mnd).
 }
}
