% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_aldersklasse}
\alias{utlede_aldersklasse}
\title{Add variable `aldersklasse` - decades}
\usage{
utlede_aldersklasse(df, var = PasientAlder)
}
\arguments{
\item{df}{data.frame where the new variable should be added.}

\item{var}{Variable name of the numerical variable that contains age as
a continuous value. Default variable name is `PasientAlder`}
}
\value{
data.frame with new variable
}
\description{
Under 18 years og above 99 are set as NA
}
\examples{
df <-  data.frame(age = c(10,20,59,60, NA, 61,69,70,99,100))
df \%>\% utlede_aldersklasse(., var = age)
}
