% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_annen_diag_variabler.R
\name{utlede_annenDiag_variabler}
\alias{utlede_annenDiag_variabler}
\alias{utlede_kar_annen_diag}
\alias{legg_til_trykk_bilde_per_kar}
\alias{legg_til_trykkmaalinger}
\title{New variables from NORIC's AnnenDiag-table}
\usage{
utlede_kar_annen_diag(df_ad)

legg_til_trykk_bilde_per_kar(df_ap, df_ad, metodeType = "FFR")

legg_til_trykkmaalinger(df_ap, df_ad)
}
\arguments{
\item{df_ad}{NORIC's \emph{AnnenDiag}-table, must contain variables
\code{ForlopsID}, \code{AvdRESH}, \code{segment}, \code{graft} and
\code{metode}}

\item{df_ap}{NORIC's \emph{AngioPCIVar}-table, must contain variables
\code{ForlopsID} and \code{AVdRESH}}

\item{metodeType}{value of variable \code{metode} from NORIC-table
\emph{AnnenDiag}. Must have one of following values : "FFR", "iFR", "IVUS"
or "OCT"}
}
\description{
Functions for creating variables based on NORIC's \emph{AnnenDiag}-table and
for merging these variables into NORIC's \emph{AngioPCIVar}-table by key
variables \code{ForlopsID} and \code{AvdRESH}.
}
\details{
The function \code{utlede_kar_annen_diag()} groups the
\code{segments} in \code{df_ad} into coronary arteries in the new variable
\code{kar}, then \code{df_ad} is returned from the function.

The function \code{legg_til_trykk_bilde_per_kar()} uses the data in
\code{df_ad} to calculate 5 new variables, where variable names prefix is
type of method (one of: "FFR", "iFR", "IVUS" or "OCT") and suffix is each
level of \code{kar} (p.ex. if metode is 'FFR' then variables are
 \code{FFR_LAD}, \code{FFR_RCA}, \code{FFR_CX} and \code{FFR_Graft}).
The variable says whether or not a measure of this method
is performed in each \code{kar}. If at least one measure is done in a
\code{kar} the value of the corresponding variable is given the value
\emph{ja}, otherwise the value is \emph{nei}. These 5 variables are then
merged into \code{df_ap}, before \code{df_ap} is returned from the function.
For procedures where no measures are done on segment-level (e.g. no entries
available in \code{df_ad}), the values for the 5 new variables are <NA>.

The function \code{legg_til_trykkmaalinger()} adds variables \code{PdPa},
\code{IMR}, \code{Pa} and \code{Pd} to \code{df_ap}-table. Values of these 4
new variables (measures of pressure) depends on the \code{df_ad}-table:
value "Ja" is attributed if the method is registered for the procedure,
value "Nei" is attributed if no if this method not is performed.
For procedures where no measures are done on segment-level (e.g. no entries
available in \code{df_ad}), the values for the 4 new variables are <NA>.
}
