% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_annen_diag_variabler.R
\name{utlede_annenDiag_variabler}
\alias{utlede_annenDiag_variabler}
\alias{utlede_kar_annen_diag}
\alias{legg_til_ffr_per_kar}
\alias{legg_til_ifr_per_kar}
\alias{legg_til_ivus_per_kar}
\alias{legg_til_oct_per_kar}
\title{New variables from NORIC's AnnenDiag-table}
\usage{
utlede_kar_annen_diag(df_ad = ad)

legg_til_ffr_per_kar(df_ap, df_ad)

legg_til_ifr_per_kar(df_ap, df_ad)

legg_til_ivus_per_kar(df_ap, df_ad)

legg_til_oct_per_kar(df_ap, df_ad)
}
\arguments{
\item{df_ad}{NORIC's \emph{AnnenDiag}-table, must contain variables
\code{ForlopsID}, \code{AVdRESH}, \code{segment}, \code{graft} and
\code{metode}}

\item{df_ap}{NORIC's \emph{AngioPCIVar}-table, must contain variables
\code{ForlopsID} and \code{AVdRESH}}
}
\description{
Functions for creating variables based on NORIC's \emph{AnnenDiag}-table and
for merging these variables into NORIC's \emph{AngioPCIVar}-table by key
variables \code{ForlopsID} and \code{AvdRESH}.
}
\details{
The function \code{utlede_kar_annen_diag()} groups the
\code{segments} in \code{df_ad} into coronary arteries in the new variable
\code{kar}, then \code{df_ad} is returned from the function.

The function \code{legg_til_ffr_per_kar()} uses the data in \code{df_ad} to
calculate 5 new variables, \code{FFR_LMS}, \code{FFR_LAD}, \code{FFR_RCA},
\code{FFR_CX} and \code{FFR_Graft}, who says whether or not a FFR measure
is performed in each \code{kar}. If at least one FFR measure is done in a
\code{kar} the value of the corresponding variable is given the value
\emph{ja}, otherwise the value is \emph{nei}. These 5 variables are then
merged into \code{df_ap}, before \code{df_ap} is returned from the function.
For procedures where no measures are done on segment-level (e.g. no entries
available in \code{df_ad}), the values for the 5 new variables are <NA>.
}
\examples{
df_ad <- data.frame(
   ForlopsID = c(1, 1, 3, 3, 5),
   AvdRESH = rep(1,5),
   segment = c("PDA/RPD (4)",
               "PLA (18)",
               "Midtre LAD (7)",
               "Ve hovedstamme (5)",
               "Ve hovedstamme (5)"),
   graft = c("Nei",
             "Arteriell",
             "Vene",
             "Nei",
             "Nei"),
   metode = c("FFR", "FFR", NA, "test", "FFR"))


df_ap <-data.frame(
   ForlopsID = c(1, 2, 3, 4, 5),
   AvdRESH = rep(1,5))


utlede_kar_annen_diag(df_ad)
legg_til_ffr_per_kar(df_ap = df_ap,
                     df_ad = df_ad)
}
