% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_ferdigstilt}
\alias{utlede_ferdigstilt}
\title{Add binary variable ("ja", "nei") for table status}
\usage{
utlede_ferdigstilt(df, var = SkjemaStatusStart, suffix = "startSkjema")
}
\arguments{
\item{df}{data.frame where variable should be added}

\item{var}{variable name of the numerical staus-variable with
values {-1,0,1}}

\item{suffix}{wanted suffix in the added variable's name:ferdigstilt_`suffix`}
}
\value{
data.frame with new variable
}
\description{
Add binary variable ("ja", "nei") for table status
}
\examples{
x <- data.frame(SkjemaStatusStart = c(-1,-1,0,0,1,-1,0,1,NA),
                SkjemastatusHovedskjema = rep(0,9),
                SkjemaStatusUtskrivelse = rep(-1,9),
                SkjemaStatusKomplikasjoner = c(rep(-1,4), rep(1,3), NA, NA))
x \%>\% utlede_ferdigstilt(., var = SkjemaStatusStart, suffix = "startSkjema")
x \%>\%
utlede_ferdigstilt(., var = SkjemaStatusStart, suffix = "startSkjema") \%>\%
utlede_ferdigstilt(., var = SkjemastatusHovedskjema,
suffix = "hovedSkjema") \%>\%
utlede_ferdigstilt(., var = SkjemaStatusKomplikasjoner,
 suffix = "komplikSkjema")
}
