% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_indikatorer.R
\name{utlede_kvalitesindikatorer}
\alias{utlede_kvalitesindikatorer}
\alias{ki_ferdigstilt_komplikasjoner}
\alias{ki_trykkmaaling_utfoert}
\alias{ki_ivus_oct_ved_stenting_lms}
\alias{ki_foreskr_blodfortynnende}
\alias{ki_foreskr_kolesterolsenkende}
\alias{ki_nstemi_utredet_innen24t}
\alias{ki_nstemi_utredet_innen72t}
\alias{ki_stemi_pci_innen120min}
\alias{ki_ak_pacemakerbehov}
\title{NORIC's quality indicators}
\usage{
ki_ferdigstilt_komplikasjoner(df_ap)

ki_trykkmaaling_utfoert(df_ap)

ki_ivus_oct_ved_stenting_lms(df_ap)

ki_foreskr_blodfortynnende(df_ap)

ki_foreskr_kolesterolsenkende(df_ap)

ki_nstemi_utredet_innen24t(df_ap)

ki_nstemi_utredet_innen72t(df_ap)

ki_stemi_pci_innen120min(df_ap)

ki_ak_pacemakerbehov(df_ak)
}
\arguments{
\item{df_ap}{NORIC's \code{AngioPCIVar}-table. Depending on indicators, must
contain some of the variables \code{SkjemaStatusKomplikasjoner},
\code{SkjemaStatusUtskrivelse}, \code{Regtype},
\code{Indikasjon}, \code{FFR}, \code{IFR},
\code{satt_inn_stent_i_LMS}, \code{TidlABC}, \code{IVUS}, \code{OCT},
\code{antall_stent_under_opphold}, ...}

\item{df_ak}{NORIC's \code{AortaklaffVar}-table. 
Depending on indicators, must contain some of the variables
 \code{AvdKompPacemaker}, \code{TypeKlaffeprotese}, ...}
}
\description{
Functions for creating two new variables for each indicator:
The first variable says whether the procedure counts (ja/nei) in the
denominator for the given indicator and the second says whether
the procedure counts in the numerator (ja/nei/NA). If the denominator has
value \emph{nei} the numerator should always be \emph{NA}. Denominator
variables names are suffixed by \emph{_data} (datagrunnlag).
}
\details{
Following variables are created for the different functions:

\code{ki_ferdigstilt_komplikasjoner()}
 \itemize{
 \item denominator \code{indik_komplik_ferdig_data} (datagrunnlag) is
 \emph{ja} when \emph{Regtype = Primaer} and 
 \code{SkjemaStatusKomplikasjoner} is \emph{-1, 0, 1}. 
 \item numerator \code{indik_komplik_ferdig} has value \emph{ja} if the
 complications form is completed (komplikasjons-skjema er ferdigstilt).}



\code{ki_trykkmaaling_utfoert()}
\itemize{
\item denominator \code{indik_trykkmaaling_data} (datagrunnlag) is \emph{ja}
when \emph{Indikasjon = Stabil koronarsykdom}.
\item numerator \code{indik_trykkmaaling} has value \emph{ja} if one of FFR,
 iFR, Pd/Pa, IMR, Pa or Pd has been performed.}

\code{ki_ivus_oct_ved_stenting_lms()}
\itemize{
\item denominator \code{indik_ivus_oct_v_stent_lms_data} (datagrunnlag) is
\emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{Indikasjon} is one of "Vitieutredning", "Uklare brystsmerter",
 "Annet", "Hjertestans uten STEMI", "Hjertesvikt/kardiomyopati",
  "Komplettering av tidligere PCI", "UAP", "NSTEMI",
  "Stabil koronarsykdom"
  \item \code{satt_inn_stent_i_LMS} = "ja"
  \item \code{TidlABC} is one of "Nei", "Ukjent", NA
  }
\item numerator \code{indik_ivus_oct_v_stent_lms} has value \emph{ja} if
IVUS and/or OCT has been performed.
}

\code{ki_foreskr_blodfortynnende()}
\itemize{
\item denominator \code{indik_blodfortynnende_data} (datagrunnlag) is
\emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{antall_stent_under_opphold} is at least 1
  \item \code{Regtype} = "Primær"
  \item \code{UtskrevetDod}  = ("Nei","NA"). Which means that  "Ja" or
  "Ukjent" are excluded
  }
\item numerator \code{indik_blodfortynnende} has value \emph{ja} if
at least two of \code{ASA}, \code{AndrePlatehemmere} or
 \code{Antikoagulantia} are prescribed and \code{SkjemaStatusUtskrivelse}
 is 1 (ferdigstilt). \code{indik_blodfortynnende} has value \emph{nei} if
at less than two of \code{ASA}, \code{AndrePlatehemmere} or
 \code{Antikoagulantia} are prescribed and \code{SkjemaStatusUtskrivelse}
 is 1 (ferdigstilt). \code{indik_blodfortynnende} has value
  \emph{ikke ferdigstilt} if  \code{SkjemaStatusUtskrivelse} is different
  from 1 (ikke ferdigstilt).
}

\code{ki_foreskr_kolesterolsenkende()}
\itemize{
\item denominator \code{indik_kolesterolsenkende_data}
(datagrunnlag) is \emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{antall_stent_under_opphold} is at least 1
  \item \code{Regtype} = "Primær"
  \item \code{UtskrevetDod}  = ("Nei", NA) . Which means that "Ja" or
  "Ukjent" are excluded
  }
\item numerator \code{indik_kolesterolsenkende} has value \emph{ja}
if \code{UtskrStatiner} is "Ja" and \code{SkjemaStatusUtskrivelse}
 is 1 (ferdigstilt). \code{indik_kolesterolsenkende} has value \emph{nei} if
\code{UtskrStatiner} is different from "Ja" and 
\code{SkjemaStatusUtskrivelse} is 1 (ferdigstilt).
 \code{indik_kolesterolsenkende} has value
  \emph{ikke ferdigstilt} if  \code{SkjemaStatusUtskrivelse} is different
  from 1 (ikke ferdigstilt).
}


\code{ki_nstemi_utredet_innen24t()}
\itemize{
\item denominator \code{indik_nstemi_angio_innen24t_data}
(datagrunnlag) is \emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{Indikasjone} is "NSTEMI"
  \item \code{Regtype} = "Primær"
  \item \code{Innkomstarsak} not "Øvrig" (NA allowed)
  \item \code{Hastegrad}  = "Akutt" or "Subakutt.
  \item \code{OverflyttetFra} neither "Annen  avdeling på sykehuset" nor
  "NA".
  }
\item numerator \code{indik_nstemi_angio_innen24t} has value \emph{ja}
if \code{ventetid_nstemi_timer} is in the interval 0-24hours,
value \emph{nei} if  \code{ventetid_nstemi_timer} is in the interval 24hours
to 14 days and value \emph{ugyldig/manglende} if time is negative, longer
than 14 days or missing. See also function
 \link[noric]{legg_til_ventetid_nstemi_timer}.
}

\code{ki_stemi_pci_innen120min()}
\itemize{
\item denominator \code{indik_stemi_pci_innen2t_data}
(datagrunnlag) is \emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{AvdRESH} is different from 106944 (AHUS Gardermoen does not
  treat STEMI patients)
  \item \code{Regtype} = "Primær"
  \item \code{Indikasjon} = "STEMI"
  \item \code{GittTrombolyse} is "Nei" or "NA"
  \item \code{Hastegrad}  = "Akutt"
  \item \code{HLRForSykehus} different from  "Ja", "Ukjent"
  \item \code{ProsedyreType} different from "Angio"
  }
\item numerator \code{indik_stemi_pci_innen2t} has value \emph{ja}
if \code{ventetid_stemi_min} is in the interval 0-120 minutes,
value \emph{nei} if  \code{ventetid_stemi_min} is in the interval 120 min
to 24h (86400 min) and value \emph{ugyldig/manglende} if time is negative,
longer than 24h,  missing or if \code{BeslutningsutlosendeEKG} has value
"Prehospitalt" and \code{ventetid_stemi_min} is 0 minutes. See also function
 \link[noric]{legg_til_ventetid_stemi_min}.
}

\code{ki_ak_pacemakerbehov()}
\itemize{
\item nevneren \code{indik_pacemakerbehov_data}
(datagrunnlag) er \emph{ja} når alle disse kriteriene er oppfylt:
\enumerate{
  \item \code{AvdRESH} tilhører sykehusene som utfører tavi :
  (HUS, UNN, Ullevål,St.Olavs, Rikshospitalet, Ahus Gardermoen)
  \item \code{TypeKlaffeprotese} ikke manglende (satt inn klaff)
  \item \code{LabKompDod} ulik \emph{ja}
  \item \code{Pacemaker}  ulik \emph{ja}
  }
\item telleren \code{indik_pacemakerbehov} har verdien \emph{ja}
dersom \code{AvdKompPacemaker} er \emph{ja} og 
\code{SkjemaStatusHovedskjema} er \emph{1} (ferdigstilte). 
 \code{indik_pacemakerbehov} har verdien \emph{nei} dersom 
 \code{AvdKompPacemaker} er \emph{nei} eller mangler, og ders
  \code{SkjemaStatusHovedskjema} er \emph{1} (ferdigstilte). 
\code{indik_pacemakerbehov} har verdien \emph{ikke ferdigstilte} dersom 
\code{SkjemaStatusHoveskjema} er \emph{-1} eller \emph{0}.
}
}
