% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_indikatorer.R
\name{utlede_kvalitesindikatorer}
\alias{utlede_kvalitesindikatorer}
\alias{ki_ferdigstilt_komplikasjoner}
\alias{ki_trykkmaaling_utfoert}
\alias{ki_ivus_oct_ved_stenting_lms}
\alias{ki_foreskr_blodfortynnende}
\alias{ki_foreskr_kolesterolsenkende}
\alias{ki_nstemi_utredet_innen24t}
\alias{ki_nstemi_utredet_innen72t}
\alias{ki_stemi_pci_innen120min}
\alias{ki_ak_pacemakerbehov}
\title{NORIC's quality indicators}
\usage{
ki_ferdigstilt_komplikasjoner(df_ap)

ki_trykkmaaling_utfoert(df_ap)

ki_ivus_oct_ved_stenting_lms(df_ap)

ki_foreskr_blodfortynnende(df_ap)

ki_foreskr_kolesterolsenkende(df_ap)

ki_nstemi_utredet_innen24t(df_ap)

ki_nstemi_utredet_innen72t(df_ap)

ki_stemi_pci_innen120min(df_ap)

ki_ak_pacemakerbehov(df_ak)
}
\arguments{
\item{df_ap}{NORIC's \code{AngioPCIVar}-table. Depending on indicators, must
contain some of the variables \code{SkjemaStatusKomplikasjoner},
\code{SkjemaStatusUtskrivelse}, \code{Regtype},
\code{Indikasjon}, \code{FFR}, \code{IFR},
\code{satt_inn_stent_i_LMS}, \code{TidlABC}, \code{IVUS}, \code{OCT},
\code{antall_stent_under_opphold}, ...}

\item{df_ak}{NORIC's \code{AortaklaffVar}-table. 
Depending on indicators, must contain some of the variables
 \code{AvdKompPacemaker}, \code{TypeKlaffeprotese}, ...}
}
\description{
Functions for creating two new variables for each indicator:
The first variable says whether the procedure counts (ja/nei) in the
denominator for the given indicator and the second says whether
the procedure counts in the numerator (ja/nei/NA). If the denominator has
value \emph{nei} the numerator should always be \emph{NA}. Denominator
variables names are suffixed by \emph{_data} (datagrunnlag).
}
\details{
Following variables are created for the different functions:

\code{ki_ferdigstilt_komplikasjoner()}
 \itemize{
 \item denominator \code{indik_komplik_ferdig_data} (datagrunnlag) is
 \emph{ja} when \emph{Regtype = Primaer} and 
 \code{SkjemaStatusKomplikasjoner} is \emph{-1, 0, 1}. 
 \item numerator \code{indik_komplik_ferdig} has value \emph{ja} if the
 complications form is completed (komplikasjons-skjema er ferdigstilt).}



\code{ki_trykkmaaling_utfoert()}
\itemize{
\item denominator \code{indik_trykkmaaling_data} (datagrunnlag) is \emph{ja}
when \emph{Indikasjon = Stabil koronarsykdom}.
\item numerator \code{indik_trykkmaaling} has value \emph{ja} if one of FFR,
 iFR, Pd/Pa, IMR, Pa or Pd has been performed.}

\code{ki_ivus_oct_ved_stenting_lms()}
\itemize{
\item denominator \code{indik_ivus_oct_v_stent_lms_data} (datagrunnlag) is
\emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{Indikasjon} is one of "Vitieutredning", "Uklare brystsmerter",
 "Annet", "Hjertestans uten STEMI", "Hjertesvikt/kardiomyopati",
  "Komplettering av tidligere PCI", "UAP", "NSTEMI",
  "Stabil koronarsykdom"
  \item \code{satt_inn_stent_i_LMS} = "ja"
  \item \code{TidlABC} is one of "Nei", "Ukjent", NA
  }
\item numerator \code{indik_ivus_oct_v_stent_lms} has value \emph{ja} if
IVUS and/or OCT has been performed.
}

\code{ki_foreskr_blodfortynnende()}
\itemize{
\item denominator \code{indik_blodfortynnende_data} (datagrunnlag) is
\emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{antall_stent_under_opphold} is at least 1
  \item \code{Regtype} = "Primær"
  \item \code{UtskrevetDod}  = ("Nei","NA"). Which means that  "Ja" or
  "Ukjent" are excluded
  }
\item numerator \code{indik_blodfortynnende} has value \emph{ja} if
at least two of \code{ASA}, \code{AndrePlatehemmere} or
 \code{Antikoagulantia} are prescribed and \code{SkjemaStatusUtskrivelse}
 is 1 (ferdigstilt). \code{indik_blodfortynnende} has value \emph{nei} if
at less than two of \code{ASA}, \code{AndrePlatehemmere} or
 \code{Antikoagulantia} are prescribed and \code{SkjemaStatusUtskrivelse}
 is 1 (ferdigstilt). \code{indik_blodfortynnende} has value
  \emph{ikke ferdigstilt} if  \code{SkjemaStatusUtskrivelse} is different
  from 1 (ikke ferdigstilt).
}

\code{ki_foreskr_kolesterolsenkende()}
\itemize{
\item denominator \code{indik_kolesterolsenkende_data}
(datagrunnlag) is \emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{antall_stent_under_opphold} is at least 1
  \item \code{Regtype} = "Primær"
  \item \code{UtskrevetDod}  = ("Nei", NA) . Which means that "Ja" or
  "Ukjent" are excluded
  }
\item numerator \code{indik_kolesterolsenkende} has value \emph{ja}
if \code{UtskrStatiner} is "Ja" and \code{SkjemaStatusUtskrivelse}
 is 1 (ferdigstilt). \code{indik_kolesterolsenkende} has value \emph{nei} if
\code{UtskrStatiner} is different from "Ja" and 
\code{SkjemaStatusUtskrivelse} is 1 (ferdigstilt).
 \code{indik_kolesterolsenkende} has value
  \emph{ikke ferdigstilt} if  \code{SkjemaStatusUtskrivelse} is different
  from 1 (ikke ferdigstilt).
}


\code{ki_nstemi_utredet_innen24t()}
\itemize{
\item denominator \code{indik_nstemi_angio_innen24t_data}
(datagrunnlag) is \emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{Indikasjone} is "NSTEMI"
  \item \code{Regtype} = "Primær"
  \item \code{Innkomstarsak} not "Øvrig" (NA allowed)
  \item \code{Hastegrad}  = "Akutt" or "Subakutt.
  \item \code{OverflyttetFra} neither "Annen  avdeling på sykehuset" nor
  "NA".
  }
\item numerator \code{indik_nstemi_angio_innen24t} has value \emph{ja}
if \code{ventetid_nstemi_timer} is in the interval 0-24hours,
value \emph{nei} if  \code{ventetid_nstemi_timer} is in the interval 24hours
to 14 days and value \emph{ugyldig/manglende} if time is negative, longer
than 14 days or missing. See also function
 \link[noric]{legg_til_ventetid_nstemi_timer}.
}

\code{ki_stemi_pci_innen120min()}
\itemize{
\item denominator \code{indik_stemi_pci_innen2t_data}
(datagrunnlag) is \emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{AvdRESH} is different from 106944 (AHUS Gardermoen does not
  treat STEMI patients)
  \item \code{Regtype} = "Primær"
  \item \code{Indikasjon} = "STEMI"
  \item \code{GittTrombolyse} is "Nei" or "NA"
  \item \code{Hastegrad}  = "Akutt"
  \item \code{HLRForSykehus} different from  "Ja", "Ukjent"
  \item \code{ProsedyreType} different from "Angio"
  }
\item numerator \code{indik_stemi_pci_innen2t} has value \emph{ja}
if \code{ventetid_stemi_min} is in the interval 0-120 minutes,
value \emph{nei} if  \code{ventetid_stemi_min} is in the interval 120 min
to 24h (86400 min) and value \emph{ugyldig/manglende} if time is negative,
longer than 24h,  missing or if \code{BeslutningsutlosendeEKG} has value
"Prehospitalt" and \code{ventetid_stemi_min} is 0 minutes. See also function
 \link[noric]{legg_til_ventetid_stemi_min}.
}

\code{ki_ak_pacemakerbehov()}
\itemize{
\item nevneren \code{indik_pacemakerbehov_data}
(datagrunnlag) er \emph{ja} når alle disse kriteriene er oppfylt:
\enumerate{
  \item \code{AvdRESH} tilhører sykehusene som utfører tavi :
  (HUS, UNN, Ullevål,St.Olavs, Rikshospitalet, Ahus Gardermoen)
  \item \code{TypeKlaffeprotese} ikke manglende (satt inn klaff)
  \item \code{LabKompDod} ulik \emph{ja}
  \item \code{Pacemaker}  ulik \emph{ja}
  }
\item telleren \code{indik_pacemakerbehov} har verdien \emph{ja}
dersom \code{AvdKompPacemaker} er \emph{ja} og 
\code{SkjemaStatusHovedskjema} er \emph{1} (ferdigstilte). 
 \code{indik_pacemakerbehov} har verdien \emph{nei} dersom 
 \code{AvdKompPacemaker} er \emph{nei} eller mangler, og ders
  \code{SkjemaStatusHovedskjema} er \emph{1} (ferdigstilte). 
\code{indik_pacemakerbehov} har verdien \emph{ikke ferdigstilte} dersom 
\code{SkjemaStatusHoveskjema} er \emph{-1} eller \emph{0}.
}
}
\examples{
 x <- data.frame(
      SkjemaStatusKomplikasjoner = c(-1, 1, 0, NA, NA, NA), 
      Regtype = c(rep("Primær", 5), NA))
 noric::ki_ferdigstilt_komplikasjoner(df_ap = x)

 x <- data.frame(
        ProsedyreDato = c(rep(as.Date("01-01-2018", format = "\%d-\%m-\%Y"), 23),
                          rep(as.Date("01-06-2016", format = "\%d-\%m-\%Y"), 2)),
        Indikasjon = c(rep("Stabil koronarsykdom", 6), NA, NA, "Annet",
                       rep("Stabil koronarsykdom", 16)),
        FFR = c(NA, "Ja", "Ja", NA, "Ja", "Ukjent", "Nei", "Ja", "Ja", rep(NA, 16)),
        IFR = c(NA, "Ja", "Nei", "Ja", "Ukjent", NA, NA, "Ja", NA,  rep(NA, 16)),
        IMR = c(rep(NA, 9),  "Ja","Ja", "Ja", "Ukjent", "Nei", "Nei", rep(NA, 10)),
        PDPA = c(rep(NA, 9), "Ja","Ja", "Ja" , "Ukjent", "Nei", "Nei", rep(NA, 10)),
        PA_Hyperemi = c(rep(NA, 9), "Ja","Ja", "Nei" , "Ukjent","Nei", "Nei", rep(NA, 10)),
        PD_Hyperemi = c(rep(NA, 9), "Ja", "Nei", "Nei", "Ukjent", "Ja", rep(NA, 11)),
        TidlABC = c("Nei", rep("Ja", 4),
                    rep(c("Nei", "Ukjent", NA_character_), 3),
                    rep(NA_character_, 11)),
        SEGMENT1 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT2 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT3 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT4 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT5 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT6 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT7 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT8 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT9 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT10 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT11 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT12 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT13 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT14 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT15 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT16 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT17 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT18 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT19 = c(rep(2, 21), 1, 1, rep(2, 2)),
        SEGMENT20 = c(rep(2, 21), 1, 1, rep(2, 2)),
        AnnenDiagHovedSpm = c("Ukjent", rep(NA, 6), "Ja", "Nei", "Nei",
                              rep("Ukjent", 3), "Ja", "Nei", "Ja",
                              rep("Ukjent", 5), rep(NA, 4)))
 noric::ki_trykkmaaling_utfoert(df_ap = x)

 x <- data.frame(
        AvdRESH = rep(1, 30),
        ForlopsID = 1:30,
        Indikasjon = rep(c("Vitieutredning",
                           "Uklare brystsmerter",
                           "Annet",
                           "Hjertestans uten STEMI",
                           "Hjertesvikt/kardiomyopati",
                           "Komplettering av tidligere PCI",
                           "UAP",
                           "NSTEMI",
                           "NoeTull!",
                           NA_character_), 3),
        TidlABC = c(rep("Ja", 10),
                    rep(c("Nei", "Ukjent", NA_character_), 6),
                    NA_character_, NA_character_),
        IVUS = rep(c("Ja", "Ja", "Nei", "Ukjent", NA_character_), 6),
        OCT = rep(c("Nei", "Ja", "Nei", "Ukjent", NA_character_), 6),
        AnnenDiagHovedSpm = c(rep("Ja", 5), rep("Nei", 5), rep("Ukjent", 2),
                              rep("Nei", 3), rep(NA_character_, 5),
                              rep(c("Ja", "Nei", "Ukjent",
                                    NA_character_, NA_character_), 2)),
        satt_inn_stent_i_LMS = c(rep("nei", 6), rep(NA_character_, 3),
                                 rep("nei", 2), rep("ja", 4), rep(NA_character_, 4),
                                 rep("ja", 6), rep("nei", 2), rep(NA_character_,3)))
 noric::ki_ivus_oct_ved_stenting_lms(df_ap = x)

 x <- data.frame(
      antall_stent_under_opphold = 1:6,
      Regtype = c("Primær", "Primær", "Sekundær", rep("Primær", 3)),
      SkjemaStatusUtskrivelse = 1,
      UtskrevetDod = c(rep("Nei", 5), "Ja"),
      ASA = c("Ja", "Nei", "Ukjent", "Ja", "Ja", "Nei"),
      AndrePlatehemmere = c(NA, "Annet", "Clopidogrel (Plavix)",
                            "Nei", "Nei", NA),
      Antikoagulantia = c(NA, "Annet", "Lixiana", "Marevan",  NA, "Exanta"),
      UtskrStatiner = c("Ja", "Ja", NA, "Nei", "Ukjent", "Ja"))
noric::ki_foreskr_blodfortynnende(df_ap = x)
noric::ki_foreskr_kolesterolsenkende(df_ap = x)

 x <- data.frame(
      Indikasjon = c("Annet", NA, "NSTEMI", "NSTEMI", "NSTEMI", "NSTEMI"),
      Regtype = c("Primær", "Primær", "Sekundær", rep("Primær", 3)),
      Innkomstarsak = rep(c("Brystesmerter", "Øvrig", "Dyspne"), 2),
      Hastegrad = rep("Subakutt", 6),
      OverflyttetFra = c("Annen  avdeling på sykehuset",
                         rep("Annet sykehus", 2),
                         rep("Omdirigert ambulanse", 3)),
      ventetid_nstemi_timer = c(0, 0, 5, 100, NA, -0.1))
noric::ki_nstemi_utredet_innen24t(df_ap = x)
noric::ki_nstemi_utredet_innen72t(df_ap = x)

 x <- data.frame(
      AvdRESH = 1:6,
      Indikasjon = c("Annet", rep("STEMI",5)),
      Regtype = c("Primær", "Primær", "Sekundær", rep("Primær", 3)),
      GittTrombolyse = rep("Nei", 6),
      Hastegrad = rep("Akutt", 6),
      HLRForSykehus = rep("Nei", 6),
      ProsedyreType = rep("Angio + PCI", 6),
      BeslutningsutlosendeEKG = rep("Prehospitalt", 6),
      ventetid_stemi_min = c(-10, 20, 110, 120, 1150, 1480))
noric::ki_stemi_pci_innen120min(df_ap = x)


x <- data.frame(
      AvdRESH = c(NA, 0,102966, 101619, 109880, rep(106944,7)), 
      TypeKlaffeprotese = c(rep("CoreValve", 2), NA, rep("CoreValve", 9)),
      LabKompDod = c("Nei", "Nei", "Nei", "Ja", NA, rep("Nei", 7)), 
      Pacemaker = c(rep("Nei", 4), "Ja", rep("Nei", 7)), 
      SkjemaStatusHovedskjema = c(rep(1, 5), 0, -1, rep(1, 5)), 
      AvdKompPacemaker = rep(c("Ja", "Nei", "Ukjent", NA), 3))
noric::ki_ak_pacemakerbehov(df_ak = x)

}
