% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_indikatorer.R
\name{utlede_kvalitesindikatorer}
\alias{utlede_kvalitesindikatorer}
\alias{ki_ferdigstilt_komplikasjoner}
\alias{ki_trykkmaaling_utfoert}
\alias{ki_ivus_oct_ved_stenting_lms}
\title{NORIC's quality indicators}
\usage{
ki_ferdigstilt_komplikasjoner(df_ap)

ki_trykkmaaling_utfoert(df_ap)

ki_ivus_oct_ved_stenting_lms(df_ap)
}
\arguments{
\item{df_ap}{NORIC's \code{AngioPCIVar}-table. Depending on indicators, must
contain some of the variables \code{SkjemaStatusKomplikasjoner},
\code{Indikasjon}, \code{FFR}, \code{IFR}, ....}
}
\description{
Functions for creating two new variables for each indicator:
The first variable says whether the procedure counts (ja/nei) in the
denominator for the given indicator and the second says whether
the procedure counts in the numerator (ja/nei/NA). If the denominator has
value \emph{nei} the numerator should always be \emph{NA}. Denominator
variables names are suffixed by \emph{_dg} (datagrunnlag).
}
\details{
Following variables are created for the different functions:

\code{ki_ferdigstilt_komplikasjoner()}
 \itemize{
 \item denominator \code{ki_komplik_ferdig_dg} (datagrunnlag) is
 \emph{ja} when \emph{Regtype = Primaer}.
 \item numerator \code{ki_komplik_ferdig} has value \emph{ja} if the
 complications form is completed (komplikasjons-skjema er ferdigstilt).}



\code{ki_trykkmaaling_utfoert()}
\itemize{
\item denominator \code{ki_trykkmaaling_dg} (datagrunnlag) is \emph{ja} when
 \emph{Indikasjon = Stabil koronarsykdom}.
\item numerator \code{ki_trykkmaaling} has value \emph{ja} if FFR and/or iFR
has been performed.}

\code{ki_ivus_oct_ved_stenting_lms()}
\itemize{
\item denominator \code{ki_ivus_oct_ved_stenting_lms_dg} (datagrunnlag) is
\emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{Indikasjon} is one of {"Vitieutredning", "Uklare brystsmerter",
 "Annet", "Hjertestans uten STEMI", "Hjertesvikt/kardiomyopati",
  "Komplettering av tidligere PCI", "UAP", "NSTEMI",
  "Stabil koronarsykdom"}
  \item \code{satt_inn_stent_i_LMS} = "ja"
  \item \code{TidlABC} is one of {"Nei", "Ukjent", NA }
  }
\item numerator \code{ki_ivus_oct_ved_stenting_lms} has value \emph{ja} if
IVUS and/or OCT has been performed.
}
}
\examples{
 x <- data.frame(
 SkjemaStatusKomplikasjoner = c(-1, 1, 0, NA, NA, NA))
 noric::ki_ferdigstilt_komplikasjoner(df_ap = x)

 x <- data.frame(
 Indikasjon = c(rep("Stabil koronarsykdom", 4), NA, "annet"),
 FFR = c("Ja", "Ja", NA, "Ukjent", "Nei", "Ja"),
 IFR = c("Ja", "Nei", "Ukjent", NA, NA, NA))
 noric::ki_trykkmaaling_utfoert(df_ap = x)

 x <- data.frame(
 Indikasjon = c(rep("Stabil koronarsykdom", 4), NA, "Annet"),
 TidlABC = rep("Nei", 6),
 satt_inn_stent_i_LMS = c(rep("ja", 4), NA, "nei"),
 IVUS = c("Ja", "Ja", NA, "Ukjent", "Nei", "Ja"),
 OCT = c("Ja", "Nei", "Ukjent", NA, NA, NA))
 noric::ki_ivus_oct_ved_stenting_lms(df_ap = x)
}
