% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_indikatorer.R
\name{utlede_kvalitesindikatorer}
\alias{utlede_kvalitesindikatorer}
\alias{ki_ferdigstilt_komplikasjoner}
\alias{ki_trykkmaaling_utfoert}
\alias{ki_ivus_oct_ved_stenting_lms}
\alias{ki_foreskr_blodfortynnende}
\alias{ki_foreskr_kolesterolsenkende}
\alias{ki_nstemi_utredet_innen24t}
\alias{ki_nstemi_utredet_innen72t}
\alias{ki_stemi_pci_innen120min}
\title{NORIC's quality indicators}
\usage{
ki_ferdigstilt_komplikasjoner(df_ap)

ki_trykkmaaling_utfoert(df_ap)

ki_ivus_oct_ved_stenting_lms(df_ap)

ki_foreskr_blodfortynnende(df_ap)

ki_foreskr_kolesterolsenkende(df_ap)

ki_nstemi_utredet_innen24t(df_ap)

ki_nstemi_utredet_innen72t(df_ap)

ki_stemi_pci_innen120min(df_ap)
}
\arguments{
\item{df_ap}{NORIC's \code{AngioPCIVar}-table. Depending on indicators, must
contain some of the variables \code{SkjemaStatusKomplikasjoner},
\code{SkjemaStatusUtskrivelse}, \code{Regtype},
\code{Indikasjon}, \code{FFR}, \code{IFR},
\code{satt_inn_stent_i_LMS}, \code{TidlABC}, \code{IVUS}, \code{OCT},
\code{antall_stent_under_opphold}, ...}
}
\description{
Functions for creating two new variables for each indicator:
The first variable says whether the procedure counts (ja/nei) in the
denominator for the given indicator and the second says whether
the procedure counts in the numerator (ja/nei/NA). If the denominator has
value \emph{nei} the numerator should always be \emph{NA}. Denominator
variables names are suffixed by \emph{_dg} (datagrunnlag).
}
\details{
Following variables are created for the different functions:

\code{ki_ferdigstilt_komplikasjoner()}
 \itemize{
 \item denominator \code{indik_komplik_ferdig_data} (datagrunnlag) is
 \emph{ja} when \emph{Regtype = Primaer}.
 \item numerator \code{indik_komplik_ferdig} has value \emph{ja} if the
 complications form is completed (komplikasjons-skjema er ferdigstilt).}



\code{ki_trykkmaaling_utfoert()}
\itemize{
\item denominator \code{indik_trykkmaaling_data} (datagrunnlag) is \emph{ja}
when \emph{Indikasjon = Stabil koronarsykdom}.
\item numerator \code{indik_trykkmaaling} has value \emph{ja} if FFR and/or
 iFR has been performed.}

\code{ki_ivus_oct_ved_stenting_lms()}
\itemize{
\item denominator \code{indik_ivus_oct_v_stent_lms_data} (datagrunnlag) is
\emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{Indikasjon} is one of "Vitieutredning", "Uklare brystsmerter",
 "Annet", "Hjertestans uten STEMI", "Hjertesvikt/kardiomyopati",
  "Komplettering av tidligere PCI", "UAP", "NSTEMI",
  "Stabil koronarsykdom"
  \item \code{satt_inn_stent_i_LMS} = "ja"
  \item \code{TidlABC} is one of "Nei", "Ukjent", NA
  }
\item numerator \code{indik_ivus_oct_v_stent_lms} has value \emph{ja} if
IVUS and/or OCT has been performed.
}

\code{ki_foreskr_blodfortynnende()}
\itemize{
\item denominator \code{ki_foreskr_blodfortynnende_dg} (datagrunnlag) is
\emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{antall_stent_under_opphold} is at least 1
  \item \code{Regtype} = "Primær"
  \item \code{SkjemaStatusUtskrivelse} is 1 (ferdigstilt)
  \item \code{UtskrevetDod}  = "Nei". Which means that "NA", "Ja" or
  "Ukjent" are excluded
  }
\item numerator \code{ki_foreskr_blodfortynnende} has value \emph{ja} if
at least two of \code{ASA}, \code{AndrePlatehemmere} or
 \code{Antikoagulantia} are prescribed.
}

\code{ki_foreskr_kolesterolsenkende()}
\itemize{
\item denominator \code{ki_foreskr_kolesterolsenkende_dg}
(datagrunnlag) is \emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{antall_stent_under_opphold} is at least 1
  \item \code{Regtype} = "Primær"
  \item \code{SkjemaStatusUtskrivelse} is 1 (ferdigstilt)
  \item \code{UtskrevetDod}  = "Nei". Which means that "NA", "Ja" or
  "Ukjent" are excluded
  }
\item numerator \code{ki_foreskr_kolesterolsenkende} has value \emph{ja}
if \code{UtskrStatiner} is "Ja".
}

\code{ki_nstemi_utredet_innen24t()}
\itemize{
\item denominator \code{ki_nstemi_utredet_innen24t_dg}
(datagrunnlag) is \emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{Indikasjone} is "NSTEMI"
  \item \code{Regtype} = "Primær"
  \item \code{Innkomstarsak} not "Øvrig" (NA allowed)
  \item \code{Hastegrad}  = "Akutt" or "Subakutt.
  \item \code{OverflyttetFra} neither "Annen  avdeling på sykehuset" nor
  "NA".
  }
\item numerator \code{ki_nstemi_utredet_innen24t} has value \emph{ja}
if \code{ventetid_nstemi_timer} is in the interval 0-24hours,
value \emph{nei} if  \code{ventetid_nstemi_timer} is in the interval 24hours
to 14 days and value \emph{ugyldig/manglende} if time is negative, longer
than 14 days or missing. See also function
 \link[noric]{legg_til_ventetid_nstemi_timer}.
}

\code{ki_stemi_pci_innen120min()}
\itemize{
\item denominator \code{ki_stemi_pci_innen120min_dg}
(datagrunnlag) is \emph{ja} when all of these conditions are fulfilled:
\enumerate{
  \item \code{AvdRESH} is different from 106944 (AHUS Gardermoen does not
  treat STEMI patients)
  \item \code{Regtype} = "Primær"
  \item \code{Indikasjon} = "STEMI"
  \item \code{GittTrombolyse} is "Nei" or "NA"
  \item \code{Hastegrad}  = "Akutt"
  \item \code{HLRForSykehus} different from  "Ja", "Ukjent"
  \item \code{ProsedyreType} different from "Angio"
  }
\item numerator \code{ki_stemi_pci_innen120min} has value \emph{ja}
if \code{ventetid_stemi_min} is in the interval 0-120 minutes,
value \emph{nei} if  \code{ventetid_stemi_min} is in the interval 120 min
to 24h (86400 min) and value \emph{ugyldig/manglende} if time is negative,
longer than 24h,  missing or if \code{BeslutningsutlosendeEKG} has value
"Prehospitalt" and \code{ventetid_stemi_min} is 0 minutes. See also function
 \link[noric]{legg_til_ventetid_stemi_min}.
}
}
\examples{
 x <- data.frame(
      SkjemaStatusKomplikasjoner = c(-1, 1, 0, NA, NA, NA))
 noric::ki_ferdigstilt_komplikasjoner(df_ap = x)

 x <- data.frame(
      Indikasjon = c(rep("Stabil koronarsykdom", 4), NA, "annet"),
      FFR = c("Ja", "Ja", NA, "Ukjent", "Nei", "Ja"),
      IFR = c("Ja", "Nei", "Ukjent", NA, NA, NA))
 noric::ki_trykkmaaling_utfoert(df_ap = x)

 x <- data.frame(
      Indikasjon = c(rep("Stabil koronarsykdom", 4), NA, "Annet"),
      TidlABC = rep("Nei", 6),
      satt_inn_stent_i_LMS = c(rep("ja", 4), NA, "nei"),
      IVUS = c("Ja", "Ja", NA, "Ukjent", "Nei", "Ja"),
      OCT = c("Ja", "Nei", "Ukjent", NA, NA, NA))
 noric::ki_ivus_oct_ved_stenting_lms(df_ap = x)

 x <- data.frame(
      antall_stent_under_opphold = 1:6,
      Regtype = c("Primær", "Primær", "Sekundær", rep("Primær", 3)),
      SkjemaStatusUtskrivelse = 1,
      UtskrevetDod = c(rep("Nei", 5), "Ja"),
      ASA = c("Ja", "Nei", "Ukjent", "Ja", "Ja", "Nei"),
      AndrePlatehemmere = c(NA, "Annet", "Clopidogrel (Plavix)",
                            "Nei", "Nei", NA),
      Antikoagulantia = c(NA, "Annet", "Lixiana", "Marevan",  NA, "Exanta"),
      UtskrStatiner = c("Ja", "Ja", NA, "Nei", "Ukjent", "Ja"))
noric::ki_foreskr_blodfortynnende(df_ap = x)
noric::ki_foreskr_kolesterolsenkende(df_ap = x)

 x <- data.frame(
      Indikasjon = c("Annet", NA, "NSTEMI", "NSTEMI", "NSTEMI", "NSTEMI"),
      Regtype = c("Primær", "Primær", "Sekundær", rep("Primær", 3)),
      Innkomstarsak = rep(c("Brystesmerter", "Øvrig", "Dyspne"), 2),
      Hastegrad = rep("Subakutt", 6),
      OverflyttetFra = c("Annen  avdeling på sykehuset",
                         rep("Annet sykehus", 2),
                         rep("Omdirigert ambulanse", 3)),
      ventetid_nstemi_timer = c(0, 0, 5, 100, NA, -0.1))
noric::ki_nstemi_utredet_innen24t(df_ap = x)
noric::ki_nstemi_utredet_innen72t(df_ap = x)

 x <- data.frame(
      AvdRESH = 1:6,
      Indikasjon = c("Annet", rep("STEMI",5)),
      Regtype = c("Primær", "Primær", "Sekundær", rep("Primær", 3)),
      GittTrombolyse = rep("Nei", 6),
      Hastegrad = rep("Akutt", 6),
      HLRForSykehus = rep("Nei", 6),
      ProsedyreType = rep("Angio + PCI", 6),
      BeslutningsutlosendeEKG = rep("Prehospitalt", 6),
      ventetid_stemi_min = c(-10, 20, 110, 120, 1150, 1480))
noric::ki_stemi_pci_innen120min(df_ap = x)
}
