% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlede_variabler.R
\name{utlede_OppholdsID}
\alias{utlede_OppholdsID}
\title{Add variable OppholdsID}
\usage{
utlede_OppholdsID(df)
}
\arguments{
\item{df}{data.frame med Angio PCI data}
}
\value{
returns \code{df} with a new column \code{OppholdsID}. For
primaerforlop \code{OppholdsID} are copied from \code{ForlopsID}, whereas
for sekundaerforlop \code{OppholdsID} are copied from
 \code{PrimaerForlopsID} aka \code{KobletForlopsID}.
\code{PrimaerForlopsID}.
}
\description{
Procedures that belongs to the same hospital-stay are given the same ID.
}
\examples{
x <- data.frame(
   Regtype = c(rep("Primær", 4), "Sekundær", "Sekundær"),
   ForlopsID = c(1, 2, 3, 4, 5, 6),
   PrimaerForlopsID = c(rep(NA, 4), 1, 3))
utlede_OppholdsID(x)
}
